code restructured
