+++ examples/service/ocean/providers/aws/rightSizingRule/create/main.go
